requirements.txt
Package           Version
----------------- ------------
cycler            0.11.0
fonttools         4.29.0
kiwisolver        1.3.2
matplotlib        3.5.1
numpy             1.21.5
packaging         21.3
Pillow            9.0.0
pip               21.3.1
pyparsing         3.0.7
python-dateutil   2.8.2
setuptools        41.2.0
six               1.16.0
torch             1.10.1+cu113
torchaudio        0.10.1+cu113
torchvision       0.11.2+cu113
typing_extensions 4.0.1
